+++ gstore/logger.py
@@ -13,9 +13,25 @@
 # You should have received a copy of the GNU General Public License
 # along with this file.  If not, see <https://www.gnu.org/licenses/>.
 
-__copyright__ = 'Copyright (C) 2020 Serghei Iakovlev'
-__version__ = '0.0.5'
-__license__ = 'GPLv3+'
-__author__ = 'Serghei Iakovlev'
-__author_email__ = 'egrep@protonmail.ch'
-__url__ = 'https://github.com/sergeyklay/gstore'
+import sys
+import logging
+
+
+def setup_logger(*args, **kwargs):
+    """
+    Setup and return the root logger object for the application.
+    """
+
+    root = logging.getLogger('gstore')
+    root.setLevel(logging.DEBUG)
+
+    handler = logging.StreamHandler(sys.stdout)
+    handler.setLevel(logging.DEBUG)
+
+    f = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+
+    formatter = logging.Formatter(f)
+    handler.setFormatter(formatter)
+    root.addHandler(handler)
+
+    return root